algorithm is up and running
